printing input type sizes

+++ 0x00-hello_world/6-size.c
@@ -9,14 +9,14 @@ int main(void)
 	char c;
 	int i;
 	long int li;
-	long long int lli;
+	long long int ll;
 	float f;
 
-	printf("Size of a char: %i byte(s).\n", (int)sizeof(c));
-	printf("Size of an int: %i byte(s).\n", (int)sizeof(i));
-	printf("Size of a long int: %i byte(s).\n", (int)sizeof(li));
-	printf("Size of a long long int: %i byte(s).\n", (int)sizeof(lli));
-	printf("Size of a float: %i byte(s).\n", (int)sizeof(f));
+	printf("Size of a char: %lli byte(s).\n", (long long int)sizeof(c));
+	printf("Size of an int: %lli byte(s).\n", (long long int)sizeof(i));
+	printf("Size of a long int: %lli byte(s).\n", (long long int)sizeof(li));
+	printf("Size of a long long int: %lli byte(s).\n", (long long int)sizeof(ll));
+	printf("Size of a float: %lli byte(s).\n", (long long int)sizeof(f));
 
 	return (0);
 }
